api refactoring: splitting functions, generic typescript
