+++ src/api/security-api.ts
@@ -1,0 +1,10 @@
+import { instance } from "./api"
+
+type GetCaptchaUrlResponseType = { url: string }
+
+export const securityAPI = {
+    getCaptchaUrl() {
+        return instance.get<GetCaptchaUrlResponseType>(`security/get-captcha-url`)
+            .then(resp => resp.data)
+    },
+}
